requirements.txt
asgiref==3.4.1
Django==3.2.9
django-appconf==1.0.5
django-filter==21.1
django-pipeline==2.0.7
django-sass-processor==1.1
Jinja2==3.0.3
libsass==0.21.0
MarkupSafe==2.0.1
Pillow==8.4.0
pytz==2021.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.2
